gateway/requirements.txt
bleak>=0.21.0
requests>=2.31.0
python-dotenv>=1.2.1